add vote for adding docker-images repo (#8)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->

I'd like to start a vote for adding docker-images repo, just like
https://github.com/oracle/docker-images.

## Solution Description
<!-- Please clearly and concisely describe your solution. -->

We can move the existing docker files in
https://github.com/oceanbase/oceanbase/tree/6797c56e2ff2f0dfb0fceec206822a46b494fbea/tools/docker/standalone
to the new repo, and gradually add missing images such as oblogproxy and
obproxy in the future.