add fork repo template and vote for goInception (#31)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->

This repo requires to fork https://github.com/hanchuanchuan/goInception
in the OceanBase origanization.

## Solution Description
<!-- Please clearly and concisely describe your solution. -->