add vote for dify repo (#33)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->

I'd like to fork https://github.com/langgenius/dify and rename it to
`dify-for-mysql`. PTAL @caifeizhi @Teingi

## Solution Description
<!-- Please clearly and concisely describe your solution. -->