add vote for dify-plugin-oceanbase repo (#46)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->

Create a repo to maintain the official dify plugin for OceanBase.


## Solution Description
<!-- Please clearly and concisely describe your solution. -->