create oceanbase-seekdb (#47)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->
Add oceanbase-seekdb repo.

## Solution Description
<!-- Please clearly and concisely describe your solution. -->

oceanbase-seekdb is a lite database based on oceanbase.

Please create a `private` repo and public it later.