add pyseekdb repo (#48)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->
Add pyseekdb repo.
pyseekdb is a python client for vector database, including OceanBase,
SeekDB service and embed SeekDB.