add lib charset tab (#51)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->

Implement a character set library for OceanBase to support better text
readability and indexing in the main project.

## Solution Description
<!-- Please clearly and concisely describe your solution. -->
add charset code table file

Co-authored-by: xiuli.qz <sanquan.qz@oceanbase.com>