add vote for langfuse repo (#53)
Added a proposal for forking the langfuse project and creating the
langfuse-bridge project, including project details and maintainers.

<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->



## Solution Description
<!-- Please clearly and concisely describe your solution. -->