feat:add vote for seekdb demo repo (#55)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->
feat:vote for add seekdb demo repo


## Solution Description
<!-- Please clearly and concisely describe your solution. -->