add vote to create ob hms client repo (#57)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->

Add hms client repo



## Solution Description
<!-- Please clearly and concisely describe your solution. -->

Add a pr to apply the ob hms client