add vote for forking ycsb repo (#65)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
add vote for forking brianfrankcooper/YCSB
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->



## Solution Description
<!-- Please clearly and concisely describe your solution. -->
add vote for forking brianfrankcooper/YCSB